django/requirements.txt
Django==3.0.8
gunicorn==20.0.4
py2neo
requests
boto3
pymongo==3.12.3
psutil
azure-mgmt-monitor
azure-identity
azure-monitor-query
azure-storage-blob
azure-mgmt-storage
azure-mgmt-compute
azure-mgmt-resource
azure-mgmt-network
msrestazure
xlwt==1.3.0
branca==0.6.0
folium==0.14.0
pygeoip
graphene-django
djangorestframework
django-filter
docker
whitenoise
neo4j
docx2pdf
aspose.slides
